+++ internal/remotes/union.go
@@ -5,12 +5,23 @@ import (
 	"fmt"
 	"log/slog"
 	"slices"
+	"sync"
 
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/model"
 )
 
 type Union struct {
 	rs []Remote
+}
+
+type mapResult[T any] struct {
+	res T
+	err *RepoAccessError
+}
+
+type joinedResult[T any] struct {
+	res  T
+	errs []*RepoAccessError
 }
 
 type RepoAccessError struct {
@@ -49,60 +60,182 @@ func NewUnion(rs ...Remote) *Union {
 }
 
 func (u *Union) Fetch(id string) (string, []byte, error, []*RepoAccessError) {
-	var errs []*RepoAccessError
-	for _, r := range u.rs {
-		id, thing, err := r.Fetch(id)
-		if err == nil {
-			return id, thing, nil, nil
-		} else {
-			if !errors.Is(err, ErrTmNotFound) {
-				errs = append(errs, NewRepoAccessError(r.Spec(), err))
-			}
-		}
+	type fetchRes struct {
+		id  string
+		b   []byte
+		err error
 	}
-	msg := fmt.Sprintf("No thing model found for %v", id)
-	slog.Default().Error(msg)
-	return "", nil, ErrTmNotFound, errs
+
+	mapper := func(r Remote) mapResult[fetchRes] {
+		fid, thing, err := r.Fetch(id)
+		res := fetchRes{id: fid, b: thing, err: err}
+		if err != nil && !errors.Is(err, ErrTmNotFound) {
+			return mapResult[fetchRes]{res: res, err: NewRepoAccessError(r.Spec(), err)}
+		}
+		return mapResult[fetchRes]{res: res, err: nil}
+	}
+	out := mapFirst[fetchRes](u.rs, mapper, func(res fetchRes) bool { return res.err == nil }, fetchRes{err: ErrTmNotFound})
+	res := <-out
+
+	if res.res.err != nil {
+		msg := fmt.Sprintf("No thing model found for %v", id)
+		slog.Default().Error(msg)
+		return "", nil, ErrTmNotFound, res.errs
+	}
+
+	return res.res.id, res.res.b, nil, res.errs
 }
 
 func (u *Union) List(search *model.SearchParams) (model.SearchResult, []*RepoAccessError) {
-	var errs []*RepoAccessError
-	res := &model.SearchResult{}
-	for _, remote := range u.rs {
-		toc, err := remote.List(search)
+	mapper := func(r Remote) mapResult[*model.SearchResult] {
+		var raErr *RepoAccessError
+		toc, err := r.List(search)
 		if err != nil {
-			errs = append(errs, NewRepoAccessError(remote.Spec(), err))
-			continue
+			raErr = NewRepoAccessError(r.Spec(), err)
 		}
-		res.Merge(&toc)
+		return mapResult[*model.SearchResult]{
+			res: &toc,
+			err: raErr,
+		}
 	}
-	return *res, errs
+
+	reducer := func(t1, t2 *model.SearchResult) *model.SearchResult {
+		t1.Merge(t2)
+		return t1
+	}
+
+	out := mapReduce[*model.SearchResult](u.rs, mapper, &model.SearchResult{}, reducer)
+	res := <-out
+	return *res.res, res.errs
 }
 
-func (u *Union) Versions(name string) ([]model.FoundVersion, []*RepoAccessError) {
+// mapReduce performs a concurrent map of remotes to mapResult[T], then reduces the results to a single joinedResult[T]
+// returns a channel which is guaranteed to receive exactly one value
+func mapReduce[T any](remotes []Remote, mapper func(r Remote) mapResult[T], identity T, reducer func(t1, t2 T) T) <-chan joinedResult[T] {
+	results := make(chan mapResult[T])
+	wg := sync.WaitGroup{}
+	wg.Add(len(remotes))
+	go func() {
+		wg.Wait()
+		close(results)
+	}()
+
+	// start mapping goroutines
+	for _, remote := range remotes {
+		go func(r Remote) {
+			results <- mapper(r)
+			wg.Done()
+		}(remote)
+	}
+
+	out := make(chan joinedResult[T])
+	go reduce(results, identity, reducer, out)
+	return out
+}
+
+// reduce reads results from ch until ch is closed and reduces them to a single joinedResult with identity as the starting value
+func reduce[T any](ch <-chan mapResult[T], identity T, reducer func(t1, t2 T) T, out chan<- joinedResult[T]) {
+	accumulator := identity
 	var errs []*RepoAccessError
-	var res []model.FoundVersion
-	for _, remote := range u.rs {
-		vers, err := remote.Versions(name)
+	for res := range ch {
+		accumulator = reducer(accumulator, res.res)
+		if res.err != nil {
+			errs = append(errs, res.err)
+		}
+	}
+	out <- joinedResult[T]{
+		res:  accumulator,
+		errs: errs,
+	}
+}
+
+// mapFirst maps all remotes concurrently and returns the first successful result
+// returns a channel which is guaranteed to receive exactly one value
+func mapFirst[T any](remotes []Remote, mapf func(r Remote) mapResult[T], isSuccess func(res T) bool, none T) <-chan joinedResult[T] {
+	results := make(chan mapResult[T])
+	out := make(chan joinedResult[T])
+
+	done := make(chan struct{})
+	wg := sync.WaitGroup{}
+	wg.Add(len(remotes))
+	go func() {
+		wg.Wait()
+		close(results)
+	}()
+
+	// start goroutines with cancellable mapping functions
+	for _, remote := range remotes {
+		go func(r Remote) {
+			select {
+			case <-done:
+			case results <- mapf(r):
+			}
+			wg.Done()
+		}(remote)
+	}
+
+	go selectFirstSuccessful(results, done, isSuccess, none, out)
+	return out
+}
+
+// selectFirstSuccessful reads results from ch until it finds the first successful with isSuccess. If no successful result found,
+// returns joinedResult with none and any accumulated errors
+func selectFirstSuccessful[T any](ch <-chan mapResult[T], done chan struct{}, isSuccess func(res T) bool, none T, out chan<- joinedResult[T]) {
+	var errs []*RepoAccessError
+	for res := range ch {
+		if isSuccess(res.res) {
+			close(done)
+			out <- joinedResult[T]{
+				res:  res.res,
+				errs: nil,
+			}
+			return
+		}
+		if res.err != nil {
+			errs = append(errs, res.err)
+		}
+	}
+	out <- joinedResult[T]{
+		res:  none,
+		errs: errs,
+	}
+}
+func (u *Union) Versions(name string) ([]model.FoundVersion, []*RepoAccessError) {
+	mapper := func(r Remote) mapResult[[]model.FoundVersion] {
+		var raErr *RepoAccessError
+		vers, err := r.Versions(name)
 		if err != nil {
 			if !errors.Is(err, ErrTmNotFound) {
-				errs = append(errs, NewRepoAccessError(remote.Spec(), err))
+				raErr = NewRepoAccessError(r.Spec(), err)
 			}
-			continue
 		}
-		res = model.MergeFoundVersions(res, vers)
+		return mapResult[[]model.FoundVersion]{
+			res: vers,
+			err: raErr,
+		}
 	}
-	return res, errs
+	reducer := func(t1, t2 []model.FoundVersion) []model.FoundVersion {
+		return model.MergeFoundVersions(t1, t2)
+	}
+	var ident []model.FoundVersion
+	out := mapReduce[[]model.FoundVersion](u.rs, mapper, ident, reducer)
+	res := <-out
+
+	return res.res, res.errs
 }
 
 func (u *Union) ListCompletions(kind string, toComplete string) []string {
-	var cs []string
-	for _, r := range u.rs {
+	mapper := func(r Remote) mapResult[[]string] {
 		rcs, err := r.ListCompletions(kind, toComplete)
-		if err == nil {
-			cs = append(cs, rcs...)
+		if err != nil {
+			rcs = nil
 		}
+		return mapResult[[]string]{res: rcs, err: nil}
 	}
+	reducer := func(r1, r2 []string) []string { return append(r1, r2...) }
+	var cs []string
+	out := mapReduce(u.rs, mapper, cs, reducer)
+	cs = (<-out).res
 	slices.Sort(cs)
 	return slices.Compact(cs)
 }